feat(docs): create a jsdocs for starIcon, socialList components
