+++ components/starIcon/StarIcon.inteface.ts
@@ -1,6 +1,21 @@
+/**
+ * Пропсы компонента StarIcon.
+ * Используется для отображения звёзд в рейтинге (например, в отзывах, оценке товаров).
+ */
 export interface StarIconProps {
-  /** Режим редактирования (влияет на tabIndex) */
+  /**
+   * Режим редактирования (влияет на tabIndex, курсор и роль).
+   * Если true, звезда становится интерактивной: получает tabIndex=0, роль button,
+   * курсор pointer и aria-label в зависимости от состояния.
+   * Если false, звезда статична: tabIndex=-1, роль img, курсор default.
+   * @default false
+   */
   isEditable?: boolean;
-  /** Состояние закрашенности звезды */
+  /**
+   * Состояние закрашенности звезды.
+   * Если true, звезда заливается чёрным цветом (#000000).
+   * Если false, звезда заливается серым цветом (#E2E2E2).
+   * @default false
+   */
   isFilled?: boolean;
 }
