+++ mocks/socials.mock.ts
@@ -1,0 +1,24 @@
+/**
+ * Интерфейс, описывающий социальную сеть.
+ */
+export interface Social {
+  /** Уникальный идентификатор социальной сети */
+  id: number;
+  /** Название социальной сети (например, "Facebook") */
+  name: string;
+  /** Путь к иконке социальной сети в папке public */
+  icon: string;
+  /** URL-адрес страницы социальной сети */
+  url: string;
+}
+
+export const socials: Social[] = [
+  {
+    id: 1, icon: '/in.svg',
+    name: 'In',
+    url: "#"
+  },
+  { id: 2, icon: '/facebook.svg', name: 'Facebook', url: '#' },
+  { id: 3, icon: '/insta.svg', name: 'Instagram', url: '#' },
+  { id: 4, icon: '/twitter.svg', name: 'Twitter', url: '#' },
+];
